\name{aggrs}
\alias{aggrs}
\title{Intermediate level function}
\description{Intermediate level function to aggregate data and create new data.frame. Required inputs are data.frame with r (correlation coefficients) and n (sample size)
}
\usage{
aggrs(r)
}
\arguments{
  \item{r}{ data.frame with r (correlation coefficients) and n (sample size)for each study.}
}
\value{ Aggregated correlations
}
\references{ Hunter, J. E., Schmidt, F. L. (2004). Methods of meta-analysis (2nd edition). Thousand Oaks, CA: Sage.
}
\author{AC Del Re & William T. Hoyt
}
\keyword{internal}

