\name{cat_sum2}
\alias{cat_sum2}
\title{Intermediate level function}
\description{Intermediate level function for categorical moderation. Adds an overall category to cat_sum1 function. 
}
\usage{
cat_sum2(meta, mod)
}
\arguments{
  \item{meta}{ data.frame with r (correlation coefficients) and n (sample size)for each study.
}
  \item{mod}{Categorical moderator variable used for moderator analysis.
}
}
\value{ Adds an overall category to cat_sum1 function. 
}
\references{ Cooper, H., Hedges, L.V., & Valentine, J.C. (2009). The handbook of research synthesis and meta-analysis (2nd edition). New York: Russell Sage Foundation. 
}
\author{ AC Del Re & William T. Hoyt
}
\keyword{internal}

