% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compile_admb2}
\alias{compile_admb2}
\title{Compile  ADMB code}
\usage{
compile_admb2(tpl_name = "srest.tpl", windose = FALSE, compile = TRUE)
}
\arguments{
\item{tpl_name}{Name of the admb code}

\item{windose}{Whether in windows (default FALSE) or not}

\item{compile}{Compile binaries from source. TRUE requires ADMB to be installed, FALSE only works on Windows}
}
\description{
The \code{plotMSY} approach uses a ADModel Builder code. ...
}

