% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqsim2.R
\name{eqsim_ggplot2}
\alias{eqsim_ggplot2}
\title{Plots of the results from eqsim}
\usage{
eqsim_ggplot2(sim, Scale = 1000, plotit = TRUE)
}
\arguments{
\item{sim}{An object returned from the function eqsim_run}

\item{Scale}{A value, the scaling on the yaxis}

\item{plotit}{Boolean, if TRUE (default) returns a plot}
}
\description{
XXX
}
\author{
Einar Hjorleifsson \email{einar.hjorleifsson@gmail.com}
}
