% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hcr_management_bio}
\alias{hcr_management_bio}
\title{Biomass-based Harvest Control Rule}
\usage{
hcr_management_bio(y, h, bio, ssb, ctr)
}
\arguments{
\item{y}{XXX}

\item{h}{XXX}

\item{bio}{vector Observed reference biomass.}

\item{ssb}{vector Observed spawning stock biomass.}

\item{ctr}{XXX}
}
\description{
The Biomass-based HCR is used in the case of the Icelandic cod
and saithe. Here it is implemented such that the TAC next year is a multiplier
of the target harvest rate and the observed reference biomass (i.e. that
includes observation errrors).

If the Btrigger is set in the rule (Btrigger > 0) then linear reductions of
fishing mortality is done relative to observed spawning stock biomass (i.e.
that includes observation errrors).
}
\note{
To do: Modify function so that buffer is not active below Btrigger and
also to a EU type TAC-constraint.
}

