% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr_management_bio2.R
\name{hcr_management_bio2}
\alias{hcr_management_bio2}
\title{iCod Biomass-based Harvest Control Rule}
\usage{
hcr_management_bio2(y, h, ctr)
}
\arguments{
\item{y}{XXX}

\item{h}{XXX}

\item{ctr}{XXX}

\item{bio}{vector Observed reference biomass.}

\item{ssb}{vector Observed spawning stock biomass.}
}
\description{
The Biomass-based HCR is used in the case of the Icelandic cod
and saithe. Here it is implemented such that the TAC next year is a multiplier
of the target harvest rate and the observed reference biomass (i.e. that 
includes observation errrors).

If the Btrigger is set in the rule (Btrigger > 0) then linear reductions of 
fishing mortality is done relative to observed spawning stock biomass (i.e.
that includes observation errrors).
}
\note{
To do: Modify function so that buffer is not active below Btrigger and
also to a EU type TAC-constraint.
}
