% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hcr_management_bio_pred}
\alias{hcr_management_bio_pred}
\title{B-based Harvest Control Rule}
\usage{
hcr_management_bio_pred(y, h, ctr)
}
\arguments{
\item{y}{XXX}

\item{h}{XXX}

\item{ctr}{Control file}
}
\description{
The B-based rule is the conventional iCod decision rule. Here it
is implemented such that the TAC next year is calculated from the observed
reference biomass.

If the Btrigger is set in the rule (Btrigger > 0) then linear reductions of
fishing mortality is done relative to observed spawning stock biomass (i.e.
that includes observation errrors).

NOTE: The model currently only works on delay=0, i.e. we are basing the advise
      on the biomass in the assessment year.

NOTE: The 10% contraint on the harvest rate not implemented yet.
}
\note{
XXX
}

