% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hcr_management_effort}
\alias{hcr_management_effort}
\title{Effort type management}
\usage{
hcr_management_effort(hrate_hat, ssb_hat, Btrigger)
}
\arguments{
\item{hrate_hat}{XXX}

\item{ssb_hat}{XXX}

\item{Btrigger}{XXX}
}
\description{
XXX
}

