% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr_management_fmort_pred.R
\name{hcr_management_fmort_pred}
\alias{hcr_management_fmort_pred}
\title{F-based Harvest Control Rule}
\usage{
hcr_management_fmort_pred(y, h, ctr)
}
\arguments{
\item{y}{XXX}

\item{h}{XXX}

\item{ctr}{Control file}
}
\description{
The F-based rule is the conventional ICES decision rule. Here it
is implemented such that the TAC next year is calculated from the true
stock in numbers based on a fishing mortality that includes observation error.

If the Btrigger is set in the rule (Btrigger > 0) then linear reductions of 
fishing mortality is done relative to observed spawning stock biomass (i.e.
that includes observation errrors).
}
\note{
XXX
}
