% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr_observation_error.R
\name{hcr_observation_error}
\alias{hcr_observation_error}
\title{Observation error model}
\usage{
hcr_observation_error(y, h, Fmult, ctr)
}
\arguments{
\item{y}{XXX}

\item{h}{XXX}

\item{Fmult}{XXX. Why is this here???}

\item{ctr}{XXX}
}
\description{
XXX
}
