% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr_operating_model.R
\name{hcr_operating_model}
\alias{hcr_operating_model}
\title{Operating model}
\usage{
hcr_operating_model(y, h, ctr, Fmult, nR = 1)
}
\arguments{
\item{y}{XXX}

\item{h}{XXX}

\item{ctr}{ctr_rec}

\item{Fmult}{XXX}

\item{nR}{Not used}
}
\description{
XXX
}
