% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hcr_recruitment_model}
\alias{hcr_recruitment_model}
\title{hcr_recruitment_model}
\usage{
hcr_recruitment_model(ssb, ctr)
}
\arguments{
\item{ssb}{The true spawning stock biomass}

\item{ctr}{The control file, containing the parameters}
}
\description{
Model to predict the recruitment
}

