% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hcr_set_dimensions}
\alias{hcr_set_dimensions}
\title{hcr_set_dimensions}
\usage{
hcr_set_dimensions(ctr)
}
\arguments{
\item{ctr}{A list containing the various dimentions as well as cv's for the
some of the parameters.}
}
\description{
Sets upp the dimensions and the error structure for the simulation.

NOTE: The current script assumes that the cv and rho in the stock are the same
as in the catches.
}

