% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hcr_set_recErrors}
\alias{hcr_set_recErrors}
\title{HCR: Setup of recruitment error structure}
\usage{
hcr_set_recErrors(d, ctr)
}
\arguments{
\item{d}{XXX}

\item{ctr}{XXX}
}
\description{
This function is modified from the original in the fishvise package.
In that package there was only one parameter estimate (a, b and cv). Here we have
different parameters for each iteration, and even different recruitment model.
}

