% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr_summarise_data.R
\name{hcr_summarise_data}
\alias{hcr_summarise_data}
\title{hcr_summarise_data}
\usage{
hcr_summarise_data(TARGET = 0.22, YEARS = c(2037:2039, ctr = ctr))
}
\arguments{
\item{TARGET}{The target harvest rate to display for the dynamic results}

\item{YEARS}{The year range over which to summarise the equilibrium results}

\item{ctr}{Control file}
}
\description{
A wrapper function for summarising the simulation results for
a given target rate
}
