% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mac_get_starting_condition}
\alias{mac_get_starting_condition}
\title{mac_get_starting_condition}
\usage{
mac_get_starting_condition(rbya, years_mean, years_cv)
}
\arguments{
\item{rbya}{data.frame}

\item{years_mean}{vector containing the years over which to take the mean}

\item{years_cv}{vector containing the years over which to take the cv}
}
\description{
Calculates mean's and cv's for weights, maturity, mortality,
partial F and M and selection pattern from rbya-objects.
}

