% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srest.R
\name{mac_srest_cat}
\alias{mac_srest_cat}
\title{Write srest.dat to disk}
\usage{
mac_srest_cat(name_stock, y1, y2, aR, aP, opt_sr_model, opt_pen, r, ssb, year)
}
\arguments{
\item{name_stock}{Name of the stock. May later down the line be used in 
table and plot outputs}

\item{y1}{First year of ssb and recruitment data}

\item{y2}{Last year of ssb and recruitment data}

\item{aR}{Recruitment age}

\item{aP}{Plus group age}

\item{opt_sr_model}{Recruitment model number (1: Ricker, 2: Beverton-Holt,
3: Segmented regression).}

\item{opt_pen}{0=no SR constraints, 1=apply SR constrain (IS THIS ONLY FOR
THE SEGMENTED REGRESSION??).}

\item{r}{A vector containing recruitment}

\item{ssb}{A vector containing spawning stock biomass}

\item{year}{A vector containg years (just used in comments)}
}
\value{
A file named \code{srest.dat} in the current working directory
}
\description{
The function is used to write srest.dat to the disk. That file
is an input to the \code{srest}-model coded in ADMB by Jose Olivera.
}
\author{
Einar Hjorleifsson
}
