\name{r_from_chi}
\alias{r_from_chi}
\title{Chi-Squared to Correlation
}
\description{Converting Chi-squared statistic with 1 df to r
}
\usage{
r_from_chi(chi.sq, n)
}
\arguments{
  \item{chi.sq}{Chi squared statistic from primary study
}
  \item{n}{Sample size in primary study
}
}
\value{Computes correlation coefficient (r)
}
\seealso{ 
\code{\link{r_from_d}},
\code{\link{r_from_d1}},
\code{\link{r_from_t}}
}
\author{AC Del Re & William T. Hoyt

Maintainer: AC Del Re \email{acdelre@gmail.com}
}
\keyword{ arith }

