\name{r_to_z}
\alias{r_to_z}
\title{Correlation to Fisher's z
}
\description{Computes Fisher's z from correlation coefficient.
}
\usage{
r_to_z(x)
}
\arguments{
  \item{x}{Correlation coefficient value.
}
}
\value{Outputs Fisher's z statistic.
}
\references{ Cooper, H., Hedges, L.V., & Valentine, J.C. (2009). \emph{The handbook of research synthesis and meta-analysis} (2nd edition). New York: Russell Sage Foundation. 
}
\author{ AC Del Re & William T. Hoyt

Maintainer: AC Del Re \email{acdelre@gmail.com}
}
\keyword{ arith }

