% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{read.conf}
\alias{read.conf}
\title{read.conf}
\usage{
read.conf(clonefile)
}
\arguments{
\item{clonefile}{XXX}
}
\description{
Reads sam configuration file. Used by read.fit function
(see below). This function originally in scr/common.R
}
