% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{read.fit}
\alias{read.fit}
\title{read.fit}
\usage{
read.fit(file, reduced = FALSE)
}
\arguments{
\item{file}{Name of file}

\item{reduced}{Boolean}
}
\description{
Seems like the big mother.

This function originally in scr/common.R
}
