% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwx2.R
\name{read.ibya}
\alias{read.ibya}
\title{read.ibya}
\usage{
read.ibya(path, Scale = 1)
}
\arguments{
\item{path}{Character containing the path to the directory}

\item{Scale}{A value}
}
\description{
read input by year and age
}
