% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.rby}
\alias{read.rby}
\title{Reads the stock summary values}
\usage{
read.rby(x, range = FALSE, Scale = 1)
}
\arguments{
\item{x}{Object from function \code{read.fit}.}

\item{range}{Boolean Not used}

\item{Scale}{A value}
}
\value{
A \code{data.frame} containing the following:
}
\description{
Reads the stock summary numbers by year from \code{sam}.
}

