% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.rbya}
\alias{read.rbya}
\title{Reads the stock in numbers and fishing mortality}
\usage{
read.rbya(x, Scale = 1)
}
\arguments{
\item{x}{Object from read.fit}

\item{Scale}{A numerical value. If 1 (default) returns the stock in numbers
as in the input file.}
}
\value{
Returns a \code{data.frame} containing the following items:
}
\description{
Reads fishing mortality and stock in numbers by year and age
from \code{sam}.
}

