\name{var_r}
\alias{var_r}
\title{Correlation Variance
}
\description{Computes variance of a correlation coefficient (r).
}
\usage{
var_r(r, n)
}
\arguments{
  \item{r}{ correlation coefficient.
}
  \item{n}{ Study sample size.
}
}
\references{ Cooper, H., Hedges, L.V., & Valentine, J.C. (2009). The handbook of research synthesis and meta-analysis (2nd edition). New York: Russell Sage Foundation. 
}
\author{ AC Del Re & William T. Hoyt

Maintainer: AC Del Re \email{acdelre@gmail.com}
}
\keyword{ arith }

