\name{var_z}
\alias{var_z}
\title{Fishers Z Variance 
}
\description{Computes the variance of a Fishers z transformation
}
\usage{
var_z(x)
}
\arguments{
  \item{x}{ Study sample size.
}
}
\value{Computes the variance of Fishers z.
}
\references{ Cooper, H., Hedges, L.V., & Valentine, J.C. (2009). The handbook of research synthesis and meta-analysis (2nd edition). New York: Russell Sage Foundation. 
}
\author{ AC Del Re & William T. Hoyt

Maintainer: AC Del Re \email{acdelre@gmail.com}
}
\keyword{ arith }

