% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ypr}
\alias{ypr}
\title{ypr}
\usage{
ypr(fmort, cW, selC, ages = 3:14, plusgroup = F, a1 = 5, a2 = 10, dW,
  ssbW, selL, M, mat, pM = 0, pF = 0, ssb = F)
}
\arguments{
\item{fmort}{A vector containing the fishing mortality over which to run the ypr}

\item{cW}{Catch weights}

\item{selC}{Selection pattern of the catch}

\item{ages}{A vector containing the age classes}

\item{plusgroup}{Boolean If FALSE (default) no plusgroup bookkeeping}

\item{a1}{A value containing lower reference age}

\item{a2}{A value containing upper reference age}

\item{dW}{Discard weights}

\item{ssbW}{Spawning stock weights}

\item{selL}{Selection pattern of the landings}

\item{M}{Natural mortality, can be a single value}

\item{mat}{A vector of proportion mature}

\item{pM}{Partial natural mortality prior to spawning}

\item{pF}{Partial fishing mortality prior to spawning}

\item{ssb}{A boolean (Not used)}
}
\description{
A yield per recruit model
}

