% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compile_admb2}
\alias{compile_admb2}
\title{Compile admb code}
\usage{
compile_admb2(tpl_name = "srest.tpl", windose = FALSE, compile = TRUE)
}
\arguments{
\item{tpl_name}{Name of the admb code}

\item{windose}{Whether in windows (default FALSE) or not}

\item{compile}{Compile binaries from source. TRUE requires ADMB to be installed, FALSE only works on Windows}
}
\value{
Nothing is returned to the R-console but leaves an executable file
in the current working directory.
}
\description{
The function compiles the \code{srest.tpl} admb-code that is bundle in the
mac package in the current working directory.

The function is a wrapper around the \code{compile_admb} function from
the \code{R2admb}-package.

The function can be used to compile \code{.tpl} codes that come with other R-packages.
}

