% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_sam_run}
\alias{get_sam_run}
\title{get_sam_run}
\usage{
get_sam_run(dir)
}
\arguments{
\item{dir}{Name of the directory on stockassessment.org}
}
\description{
Fetches the results from the run directory on stockassessment.org
and stored it locally.
}
\note{
Currently the directory structure is also copied, see datadisk/stock...
}

