% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr_set_starting_conditions.R
\name{hcr_set_starting_conditions}
\alias{hcr_set_starting_conditions}
\title{HCR: Set starting condition for stock}
\usage{
hcr_set_starting_conditions(dat_y1, d, ctr)
}
\arguments{
\item{dat_y1}{XXX}

\item{d}{XXX}

\item{ctr}{XXX}
}
\description{
XXX
}
