% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hcr_set_wgtErrors}
\alias{hcr_set_wgtErrors}
\title{hcr_set_wgtErrors}
\usage{
hcr_set_wgtErrors(d, ctr)
}
\arguments{
\item{d}{XXX}

\item{ctr}{A list with control values, here use the weight cv and weight rho.}
}
\description{
Sets up the weight error structure
}

