% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mac.R
\docType{package}
\name{mac-package}
\alias{mac-package}
\alias{mac}
\title{Title mac package}
\description{
HCR simulation of mackerel
}
\details{
Version 0.0.1: 19. June 2014
}
\seealso{
link to other documents
}
\author{
Einar Hjorleifsson
}
