% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mac_run_the_stuff}
\alias{mac_run_the_stuff}
\title{mac_run_the_stuff}
\usage{
mac_run_the_stuff(dat, ctr)
}
\arguments{
\item{dat}{A data.frame containing the mean values}

\item{ctr}{The control file}
}
\description{
A wrapper function for running the mackerel simulation
}

