% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mac_run_the_stuff2.R
\name{mac_run_the_stuff2}
\alias{mac_run_the_stuff2}
\title{mac_run_the_stuff2}
\usage{
mac_run_the_stuff2(dat, ctr)
}
\arguments{
\item{dat}{A data.frame containing the mean values}

\item{ctr}{The control file}
}
\description{
A wrapper function for running the mackerel simulation. Here
use the function \code{hcr_management_fmort_pred} function.
}
