% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.conf}
\alias{read.conf}
\title{read.conf}
\usage{
read.conf(clonefile)
}
\arguments{
\item{clonefile}{XXX}
}
\description{
Reads sam configuration file. Used by read.fit function
(see below). This function originally in scr/common.R
}

