% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwx2.R
\name{read.lowestoft}
\alias{read.lowestoft}
\title{read.lowestoft}
\usage{
read.lowestoft(filename, val.name, Format = "List")
}
\arguments{
\item{filename}{Filename}

\item{val.name}{Character Not used}

\item{Format}{Character, "List" or "Wide"}
}
\description{
Read lowestoft files
}
