% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.rby}
\alias{read.rby}
\title{read.rby}
\usage{
read.rby(x, range = FALSE, Scale = 1)
}
\arguments{
\item{x}{Object from read.fit}

\item{range}{Boolean Not used}

\item{Scale}{A value}
}
\description{
read fishing mortality and stock in numbers from sam
}

