\name{var_z}
\alias{var_z}
\title{Fisher's z Variance 
}
\description{Computes the variance of a Fisher's z transformation
}
\usage{
var_z(n)
}
\arguments{
  \item{n}{ Study sample size.
}
}
\value{Computes the variance of Fisher's z.
}
\references{ Cooper, H., Hedges, L.V., & Valentine, J.C. (2009). \emph{The handbook of research synthesis and meta-analysis} (2nd edition). New York: Russell Sage Foundation. 
}
\author{ AC Del Re & William T. Hoyt

Maintainer: AC Del Re \email{acdelre@gmail.com}
}
\keyword{ arith }

